+++ services/madoc-ts/src/frontend/admin/query-config.ts
@@ -3,7 +3,7 @@ import { ReactQueryConfig } from 'react-query';
 export const queryConfig: ReactQueryConfig = {
   queries: {
     staleTime: 0,
-    cacheTime: 0,
-    refetchOnMount: true,
+    cacheTime: 1000 * 60 * 60,
+    refetchOnMount: false,
   },
 };