+++ services/madoc-ts/src/frontend/shared/atlas/ViewReadOnlyAnnotation.tsx
@@ -6,7 +6,7 @@ import { ReadOnlyAnnotation } from '../hooks/use-read-only-annotations';
 // Avoid bad types from Atlas.
 const CustomHTMLPortal = HTMLPortal as any;
 
-export function ViewReadOnlyAnnotation({ style, ...props }: ReadOnlyAnnotation) {
+export function ViewReadOnlyAnnotation({ id, style, target, ...props }: ReadOnlyAnnotation) {
   const [isOpen, setIsOpen] = useState(false);
   const [isHover, setIsHover] = useState(false);
   const { interactive, hotspot, hotspotSize, hidden, ...allStyles } = style;
@@ -14,33 +14,50 @@ export function ViewReadOnlyAnnotation({ style, ...props }: ReadOnlyAnnotation) 
   const styleToApply = useMemo(() => {
     if (hotspot) {
       const { ':hover': hoverStyles, ':active': activeStyles, ...styles } = allStyles;
-      return isOpen ? mergeStyles(styles, activeStyles) : isHover ? mergeStyles(styles, hoverStyles) : styles;
+      return isOpen ? activeStyles : isHover ? hoverStyles : styles;
     }
     return allStyles;
   }, [allStyles, hotspot, isHover, isOpen]);
 
-  if (hotspot) {
-    return (
-      <>
-        <box interactive={interactive} {...props} style={styleToApply} />
-        <CustomHTMLPortal
-          relativeStyle
-          relativeSize={false}
-          interactive
-          target={props.target}
-          style={{ overflow: 'visible' }}
-        >
-          <HotSpot
-            $size={hotspotSize || 'md'}
-            $active={isOpen}
-            onClick={() => setIsOpen(o => !o)}
-            onMouseEnter={() => setIsHover(true)}
-            onMouseLeave={() => setIsHover(false)}
+  return (
+    <world-object id="readonly-" x={target.x} y={target.y} width={target.width} height={target.height}>
+      {hotspot ? (
+        <>
+          <box
+            id={`hotspot/${id}`}
+            html
+            interactive={interactive}
+            {...props}
+            style={styleToApply}
+            target={{ x: 0, y: 0, width: target.width, height: target.height }}
           />
-        </CustomHTMLPortal>
-      </>
-    );
-  }
-
-  return <box interactive={interactive} {...props} style={styleToApply} />;
+          <CustomHTMLPortal
+            id={`portal/${id}`}
+            relativeStyle
+            relativeSize={false}
+            interactive
+            target={{ x: 0, y: 0, width: target.width, height: target.height }}
+            style={{ overflow: 'visible' }}
+          >
+            <HotSpot
+              $size={hotspotSize || 'lg'}
+              $active={isOpen}
+              onClick={() => setIsOpen(o => !o)}
+              onMouseEnter={() => setIsHover(true)}
+              onMouseLeave={() => setIsHover(false)}
+            />
+          </CustomHTMLPortal>
+        </>
+      ) : (
+        <box
+          id={`non-hotspot/${id}`}
+          html
+          interactive={interactive}
+          {...props}
+          style={styleToApply}
+          target={{ x: 0, y: 0, width: target.width, height: target.height }}
+        />
+      )}
+    </world-object>
+  );
 }